+++ geopandas/tests/test_geodataframe.py
@@ -2,6 +2,7 @@ import json
 import os
 import shutil
 import tempfile
+from enum import Enum
 
 import numpy as np
 import pandas as pd
@@ -1488,6 +1489,19 @@ class TestConstructor:
         expected = pd.concat([df, other_df], axis=1)
         df[other_df.columns] = other_df
         assert_geodataframe_equal(df, expected)
+
+    def test_geometry_colname_enum(self):
+        # ensure that other classes to geometry arg in GeoDataFrame
+        # with `name` attribute are not assumed to be (Geo)Series
+        class Fruit(Enum):
+            apple = 1
+            pear = 2
+
+        df = pd.DataFrame(
+            {Fruit.apple: [1, 2], Fruit.pear: GeoSeries.from_xy([1, 2], [3, 4])}
+        )
+        res = GeoDataFrame(df, geometry=Fruit.pear)
+        assert res.active_geometry_name == Fruit.pear
 
 
 @pytest.mark.skipif(not compat.HAS_PYPROJ, reason="pyproj not available")